+++ include/seqan3/alignment/scoring/scoring_scheme_base.hpp
@@ -19,7 +19,6 @@
 #include <seqan3/core/concept/cereal.hpp>
 #include <seqan3/core/concept/core_language.hpp>
 #include <seqan3/core/detail/strong_type.hpp>
-#include <seqan3/range/shortcuts.hpp>
 #include <seqan3/std/algorithm>
 
 #if SEQAN3_WITH_CEREAL