+++ include/seqan3/core/type_traits/range.hpp
@@ -18,7 +18,6 @@
 #include <seqan3/core/type_traits/pre.hpp>
 #include <seqan3/core/type_traits/basic.hpp>
 #include <seqan3/core/type_traits/iterator.hpp>
-#include <seqan3/range/shortcuts.hpp>
 #include <seqan3/std/ranges>
 #include <seqan3/std/iterator>
 
@@ -149,7 +148,7 @@ template <std::ranges::sized_range rng_t>
 struct size_type<rng_t>
 {
     //!\brief Return the size_type as returned by the size function.
-    using type = decltype(size(std::declval<rng_t &>()));
+    using type = decltype(std::ranges::size(std::declval<rng_t &>()));
 };
 
 // ----------------------------------------------------------------------------