+++ include/seqan3/range/views/persist.hpp
@@ -19,7 +19,6 @@
 #include <seqan3/core/type_traits/transformation_trait_or.hpp>
 #include <seqan3/io/exception.hpp>
 #include <seqan3/range/concept.hpp>
-#include <seqan3/range/shortcuts.hpp>
 #include <seqan3/range/container/concept.hpp>
 #include <seqan3/range/views/detail.hpp>
 #include <seqan3/std/algorithm>
@@ -107,7 +106,7 @@ public:
      */
     const_iterator begin() const noexcept
     {
-        return seqan3::begin(*urange);
+        return std::ranges::begin(*urange);
     }
 
     //!\copydoc begin()
@@ -131,7 +130,7 @@ public:
      */
     auto end() const noexcept
     {
-        return seqan3::end(*urange);
+        return std::ranges::end(*urange);
     }
 
     //!\copydoc end()